refactoring: reflect separating utils.ts
